added delete method to courses

+++ src/main/java/com/devhamzat/student_management_system/entity/Course.java
@@ -21,7 +21,7 @@ public class Course {
     @Column(name = "id",
             updatable = false
     )
-    private Integer id;
+    private Long id;
     @Column(name = "course_name",
             nullable = false
     )
@@ -53,7 +53,7 @@ public class Course {
 
     private CourseType courseType;
 
-    public Course(Integer id,
+    public Course(Long id,
                   String courseName,
                   String courseCode,
                   Integer courseUnit,
@@ -70,11 +70,11 @@ public class Course {
     public Course() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
